added functionality to homepage
