+++ src/components/Feed/Feed.css
@@ -1,7 +1,7 @@
 .feed {
     flex: 0.4;
     border-right: 1px solid var(--bitter-background);
-    min-width: fit-content; 
+    min-width: fit-content;
     overflow-y: scroll;
 }
 
@@ -23,9 +23,13 @@
     z-index: 100;
     border: 1px solid var(--bitter-background);
     padding: 15px 20px;
-}   
+}
 
-.feed__header > h2 {
-    font-size: 20px;   
+.feed__header>h2 {
+    font-size: 20px;
     font-weight: 800;
+}
+
+.tweetsContainer {
+    list-style: none;
 }
