+++ src/components/Post/Post.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import './Post.css';
 import { Avatar } from "@mui/material";
 import { Verified } from "@mui/icons-material";
@@ -6,46 +6,79 @@ import { ChatBubbleOutline } from "@mui/icons-material";
 import { Repeat } from "@mui/icons-material";
 import { FavoriteBorder } from "@mui/icons-material";
 import { Publish } from "@mui/icons-material";
+import { Delete } from "@mui/icons-material";
+import CommentList from "../commentList/CommentList";
+import CommentForm from "../commentForm/CommentForm";
 
 function Post({
     displayName,
     username,
     verified,
-    text,
+    // text
     image,
-    avatar
+    avatar,
+    profile,
+    user,
+    text,
+    deleteTweet,
+    id,
+    comment,
+    tweet,
+    setComment,
+    createComment,
+    getAllComments,
+    deleteComment,
+    editComment,
+    comments
 }) {
+
+    const [toggle, setToggle] = useState(false)
+
     return (
         <>
-        <div className="post">
-            <div className="post__avatar">
-            <Avatar src="https://mystickermania.com/cdn/stickers/90-512x512.png" />
-            </div>
-            <div className="post__body">
-                <div className="post__header">
-                    <div className="post__headerText">
-                        <h3>
-                            Bitter Team {" "}
-                             <span className="post__headerSpecial">
-                                <Verified className="post__badge"></Verified>
-                            </span>
-                        </h3>
-                    </div>
-                    <div className="post__headerDescripton">
-                        <p>Bitter Team is the G.O.A.T!</p>
-                    </div>
+            <div className="post">
+                <div className="post__avatar">
+                    <Avatar src="https://mystickermania.com/cdn/stickers/90-512x512.png" />
                 </div>
-                <img src="https://media.tenor.com/TC9xkKYp6wIAAAAd/goat.gif" alt="goat"></img>
+                <div className="post__body">
+                    <div className="post__header">
+                        <div className="post__headerText">
+                            <h3>
+                                {user.username}
+                                <span className="post__headerSpecial">
+                                    <Verified className="post__badge"></Verified>
+                                </span>
+                            </h3>
+                        </div>
+                        <div className="post__headerDescripton">
+                            <p>{text}</p>
+                        </div>
+                    </div>
+                    {/* <img src="https://media.tenor.com/TC9xkKYp6wIAAAAd/goat.gif" alt="goat"></img> */}
+                </div>
+
+                {toggle === true ?
+                    <CommentList
+                        comment={comment}
+                        tweet={tweet}
+                        setComment={setComment}
+                        createComment={createComment}
+                        id={id}
+                        user={user}
+                        deleteTweet={deleteTweet}
+                        comments={comments}
+                        getAllComments={getAllComments}
+                        deleteComment={deleteComment}
+                        editComment={editComment}
+                    /> : ''}
             </div>
-            
-        </div>
-        <div className="post__footer">
-        <ChatBubbleOutline fontSize="small" />
-        <Repeat fontSize="small" />
-        <FavoriteBorder fontSize="small" />
-        <Publish fontSize="small" />
-    </div>
-    </>
+            <div className="post__footer">
+                <ChatBubbleOutline fontSize="small" onClick={() => setToggle(!toggle)} />
+                <Repeat fontSize="small" />
+                <FavoriteBorder fontSize="small" />
+                <Delete fontSize="small" onClick={() => deleteTweet(id)} />
+            </div>
+        </>
     )
 }
