+++ src/components/commentList/CommentList.js
@@ -25,8 +25,8 @@ export default function CommentList({
                 {tweet.comments.length ? tweet.comments.map((item) => (
                     <li key={item._id}>
                         <h1>{item.text}</h1>
-                        <button onClick={() => deleteComment((id, item._id))}>Delete</button>
-                        <button onClick={() => editComment((id, item._id))}>Edit</button>
+                        <button onClick={() => deleteComment(id, item._id)}>Delete</button>
+                        <button onClick={() => editComment(id, item._id)}>Edit</button>
                     </li>
                 )) : <> Be the first to comment</>}
             </ul>
